sqlite delete: remove extra punctuation
When deleting a sqlite database the prompt currently looks like:

[spin(HEAD)] $ spin cloud sqlite delete nonsense
The action is irreversible. Please type "nonsense" for confirmation.: {cursor here}

The full-stop followed by the colon is a little weird, so drop the
full-stop from the prompt to get:

[spin(HEAD)] $ spin cloud sqlite delete nonsense
The action is irreversible. Please type "nonsense" for confirmation: {cursor here}

Signed-off-by: Danielle Lancashire <dani@builds.terrible.systems>